// src/app/models/reforma.models.ts
import { Usuario } from "./usuarios.models";


export class Reforma {
    _id: string
    requirente: string;
    numeroModificacion: string;
    fechapresupuestaria: string;
    tipoGasto: string;
    tipoModificacion: string;
    montoSolicitado: string;
    justificacion: string;
    resolucion: string;
    modificacionPresupuestaria: Array<any>
    reprogramacionFinaciera: Array<any>
    usuario: Usuario;

    requisito1: string;
    requisito2: string;
    requisito3: string;
    requisito4: string;
    descripcion: string;
    
    analisisInformeRequisito: string;
    analisisInformeRequisito2: string;
    analisisInformeRequisito3: string;
    analisisInformeRequisito4: string;
    analisisInformeRequisitoDescripcion: string;

    analistaTecnico1: string;
    analistaTecnico2: string;
    analistaTecnico3: string;
    analistaTecnicoDescripcion: string;

    dicatemPresupuestaria1: string;
    dicatemPresupuestaria2: string;
    dicatemPresupuestariaDescripcion: string;

    firmaResponsabilidad: string;
    firmaResponsabilidad1: string;
    firmaResponsabilidad2: string;
    firmaResponsabilidad3: string;
    firmaResponsabilidad4: string;
    firmaResponsabilidad5: string;
    firmaResponsabilidad6: string;
    firmaResponsabilidad7: string;
    firmaResponsabilidad8: string;

    firmaAnalisisPoa: string;
    firmaAnalisisPoa1: string;
    firmaAnalisisPoa2: string;

    firmaTecnico: string;
    firmaTecnico1: string;
    firmaTecnico2: string;
    firmaTecnico3: string;
    firmaTecnico4: string;
    firmaTecnico5: string;

    firmaDicatamen: string;
    firmaDicatamen1: string;
    firmaDicatamen2: string;
    firmaDicatamen3: string;
    firmaDicatamen4: string;
    firmaDicatamen5: string;
    firmaDicatamen6: string;
    firmaDicatamen7: string;
    firmaDicatamen8: string;

    firmaAprobacion: string;
    firmaAprobacion1: string;
    firmaAprobacion2: string;
    firmaAprobacion3: string;
    firmaAprobacion4: string;
    firmaAprobacion5: string;
    firmaAprobacion6: string;
    firmaAprobacion7: string;
    firmaAprobacion8: string;

    urlPdf: string;

}